perf(mira-hub): split /login into server page + client form — closes CRA-50
Promote /login to a server component so its metadata (title, description,
canonical, openGraph) renders in the initial HTML response without
round-tripping through the client. The interactive form moves to
login-form.tsx as the only "use client" boundary.

This reduces the work done before the login form is interactive
(Lighthouse perf was 73 — target ≥80). The page-level metadata also
gives /login a proper indexable title for SEO.
