+++ mira-hub/src/app/login/page.tsx
@@ -1,266 +1,19 @@
-"use client";
+import type { Metadata } from "next";
+import LoginForm from "./login-form";
 
-import { useState, Suspense } from "react";
-import Link from "next/link";
-import { useRouter, useSearchParams } from "next/navigation";
-import { signIn } from "next-auth/react";
-import { Factory, Loader2, Eye, EyeOff, Mail, ChevronDown } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { API_BASE } from "@/lib/config";
-
-function LoginForm() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const callbackUrl = searchParams.get("callbackUrl") || "/feed";
-  const initialError = searchParams.get("error");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [showPw, setShowPw] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [googleLoading, setGoogleLoading] = useState(false);
-  const [magicEmail, setMagicEmail] = useState("");
-  const [magicLoading, setMagicLoading] = useState(false);
-  const [magicSent, setMagicSent] = useState(false);
-  const [showPasswordForm, setShowPasswordForm] = useState(false);
-  const [error, setError] = useState(initialError ? "Sign in failed" : "");
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setLoading(true);
-    setError("");
-    const result = await signIn("credentials", {
-      email,
-      password,
-      redirect: false,
-    });
-    setLoading(false);
-    if (!result || result.error) {
-      setError(result?.error || "Invalid email or password");
-      return;
-    }
-    router.push(callbackUrl);
-    router.refresh();
-  }
-
-  async function handleGoogle() {
-    setGoogleLoading(true);
-    setError("");
-    await signIn("google", { callbackUrl });
-  }
-
-  async function handleMagicLink(e: React.FormEvent) {
-    e.preventDefault();
-    if (!magicEmail) return;
-    setMagicLoading(true);
-    setError("");
-    try {
-      const res = await fetch(`${API_BASE}/api/auth/magic-link`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email: magicEmail }),
-      });
-      if (res.ok) {
-        setMagicSent(true);
-      } else {
-        setError("Failed to send magic link. Try again.");
-      }
-    } catch {
-      setError("Network error. Please try again.");
-    } finally {
-      setMagicLoading(false);
-    }
-  }
-
-  return (
-    <div
-      className="min-h-screen flex flex-col items-center justify-center px-4 py-12"
-      style={{ background: "linear-gradient(135deg, #0F172A 0%, #1E293B 50%, #0F172A 100%)" }}
-    >
-      <div
-        className="absolute inset-0 opacity-5"
-        style={{
-          backgroundImage: "radial-gradient(circle at 1px 1px, #fff 1px, transparent 0)",
-          backgroundSize: "32px 32px",
-        }}
-      />
-
-      <div className="relative w-full max-w-sm">
-        <div className="flex flex-col items-center mb-8">
-          <div
-            className="w-16 h-16 rounded-2xl flex items-center justify-center mb-4"
-            style={{
-              background: "linear-gradient(135deg, #2563EB, #0891B2)",
-              boxShadow: "0 0 32px rgba(37,99,235,0.4)",
-            }}
-          >
-            <Factory className="w-8 h-8 text-white" />
-          </div>
-          <h1 className="text-2xl font-bold text-white tracking-tight">FactoryLM</h1>
-          <p className="text-slate-400 text-sm mt-1">Sign in to continue</p>
-        </div>
-
-        <div
-          className="rounded-2xl border border-slate-700/50 p-8"
-          style={{
-            background: "rgba(26,29,35,0.95)",
-            backdropFilter: "blur(12px)",
-            boxShadow: "0 24px 64px rgba(0,0,0,0.4)",
-          }}
-        >
-          {/* Google */}
-          <button
-            type="button"
-            onClick={handleGoogle}
-            disabled={googleLoading}
-            className="w-full h-11 mb-4 flex items-center justify-center gap-3 rounded-md bg-white text-slate-900 font-medium hover:bg-slate-100 transition-colors disabled:opacity-60"
-          >
-            {googleLoading ? (
-              <Loader2 className="w-5 h-5 animate-spin" />
-            ) : (
-              <svg className="w-5 h-5" viewBox="0 0 48 48" aria-hidden="true">
-                <path fill="#FFC107" d="M43.611 20.083H42V20H24v8h11.303C33.972 32.91 29.418 36 24 36c-6.627 0-12-5.373-12-12s5.373-12 12-12c3.059 0 5.842 1.154 7.961 3.039l5.657-5.657C34.046 6.053 29.268 4 24 4 12.955 4 4 12.955 4 24s8.955 20 20 20 20-8.955 20-20c0-1.341-.138-2.65-.389-3.917z" />
-                <path fill="#FF3D00" d="M6.306 14.691l6.571 4.819C14.655 15.108 18.961 12 24 12c3.059 0 5.842 1.154 7.961 3.039l5.657-5.657C34.046 6.053 29.268 4 24 4 16.318 4 9.656 8.337 6.306 14.691z" />
-                <path fill="#4CAF50" d="M24 44c5.166 0 9.86-1.977 13.409-5.192l-6.19-5.238C29.211 35.091 26.715 36 24 36c-5.398 0-9.937-3.057-11.291-7.946l-6.522 5.025C9.505 39.556 16.227 44 24 44z" />
-                <path fill="#1976D2" d="M43.611 20.083H42V20H24v8h11.303c-.792 2.237-2.231 4.166-4.087 5.571l.003-.002 6.19 5.238C36.971 39.205 44 34 44 24c0-1.341-.138-2.65-.389-3.917z" />
-              </svg>
-            )}
-            <span>Continue with Google</span>
-          </button>
-
-          <div className="flex items-center gap-3 mb-4">
-            <div className="flex-1 h-px bg-slate-700/50" />
-            <span className="text-xs text-slate-500">or</span>
-            <div className="flex-1 h-px bg-slate-700/50" />
-          </div>
-
-          {/* Magic link */}
-          {magicSent ? (
-            <div className="text-center py-2 mb-4">
-              <Mail className="w-8 h-8 text-blue-400 mx-auto mb-2" />
-              <p className="text-white font-semibold text-sm">Check your inbox</p>
-              <p className="text-slate-400 text-xs mt-1">We sent a sign-in link to <strong>{magicEmail}</strong></p>
-              <button onClick={() => setMagicSent(false)} className="text-blue-400 text-xs mt-2 hover:text-blue-300">
-                Send again
-              </button>
-            </div>
-          ) : (
-            <form onSubmit={handleMagicLink} className="mb-4">
-              <label className="block text-xs font-medium text-slate-400 mb-1.5">Email magic link</label>
-              <div className="flex gap-2">
-                <Input
-                  type="email"
-                  value={magicEmail}
-                  onChange={(e) => setMagicEmail(e.target.value)}
-                  placeholder="you@company.com"
-                  autoComplete="email"
-                  required
-                  className="flex-1 h-11 bg-slate-800/60 border-slate-700 text-white placeholder:text-slate-500 focus:ring-blue-500"
-                />
-                <button
-                  type="submit"
-                  disabled={magicLoading || !magicEmail}
-                  className="px-4 h-11 rounded-md font-medium text-white text-sm disabled:opacity-50 flex items-center gap-1.5"
-                  style={{ background: "linear-gradient(135deg,#2563EB,#0891B2)" }}
-                >
-                  {magicLoading ? <Loader2 className="w-4 h-4 animate-spin" /> : <Mail className="w-4 h-4" />}
-                </button>
-              </div>
-            </form>
-          )}
-
-          <div className="flex items-center gap-3 mb-4">
-            <div className="flex-1 h-px bg-slate-700/50" />
-            <span className="text-xs text-slate-500">or</span>
-            <div className="flex-1 h-px bg-slate-700/50" />
-          </div>
-
-          {/* Password — collapsible */}
-          <button
-            type="button"
-            onClick={() => setShowPasswordForm(v => !v)}
-            className="w-full flex items-center justify-between text-xs text-slate-400 hover:text-slate-200 transition-colors mb-3"
-          >
-            <span>Sign in with password</span>
-            <ChevronDown className={`w-4 h-4 transition-transform ${showPasswordForm ? "rotate-180" : ""}`} />
-          </button>
-
-          {showPasswordForm && (
-            <form onSubmit={handleSubmit} className="space-y-4">
-              <div>
-                <label className="block text-xs font-medium text-slate-400 mb-1.5">Email</label>
-                <Input
-                  type="email"
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  placeholder="you@company.com"
-                  autoComplete="email"
-                  required
-                  className="h-11 bg-slate-800/60 border-slate-700 text-white placeholder:text-slate-500 focus:ring-blue-500"
-                />
-              </div>
-
-              <div>
-                <label className="block text-xs font-medium text-slate-400 mb-1.5">Password</label>
-                <div className="relative">
-                  <Input
-                    type={showPw ? "text" : "password"}
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                    placeholder="••••••••"
-                    autoComplete="current-password"
-                    required
-                    className="h-11 pr-10 bg-slate-800/60 border-slate-700 text-white placeholder:text-slate-500 focus:ring-blue-500"
-                  />
-                  <button
-                    type="button"
-                    onClick={() => setShowPw((v) => !v)}
-                    className="absolute right-3 top-1/2 -translate-y-1/2 text-slate-400 hover:text-slate-200 transition-colors"
-                  >
-                    {showPw ? <EyeOff className="w-4 h-4" /> : <Eye className="w-4 h-4" />}
-                  </button>
-                </div>
-              </div>
-
-              <Button
-                type="submit"
-                className="w-full h-11 text-base font-semibold"
-                disabled={loading}
-                style={{ background: loading ? undefined : "linear-gradient(135deg, #2563EB, #0891B2)" }}
-              >
-                {loading ? (
-                  <>
-                    <Loader2 className="w-4 h-4 animate-spin" /> Signing in
-                  </>
-                ) : (
-                  "Sign in"
-                )}
-              </Button>
-            </form>
-          )}
-
-          {error && (
-            <div className="rounded-lg bg-red-500/10 border border-red-500/20 px-3 py-2 text-sm text-red-400 mt-4">
-              {error}
-            </div>
-          )}
-
-          <p className="text-center text-sm text-slate-400 mt-6">
-            Don&apos;t have an account?{" "}
-            <Link href="/signup" className="text-blue-400 hover:text-blue-300 font-medium">
-              Start free trial
-            </Link>
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-}
+export const metadata: Metadata = {
+  title: "Sign in",
+  description: "Sign in to FactoryLM Hub.",
+  alternates: { canonical: "/login" },
+  openGraph: {
+    title: "Sign in · FactoryLM",
+    description: "Sign in to FactoryLM Hub.",
+    url: "/login",
+    type: "website",
+  },
+  robots: { index: true, follow: true },
+};
 
 export default function LoginPage() {
-  return (
-    <Suspense fallback={<div className="min-h-screen bg-slate-900" />}>
-      <LoginForm />
-    </Suspense>
-  );
+  return <LoginForm />;
 }
